# src/generated/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blawby API",
    "version": "1.0.0",
    "description": "Blawby Law Practice Management API with Stripe integration",
    "contact": {
      "name": "Blawby Support",
      "email": "support@blawby.com"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/practice/details/{organizationId}": {
      "get": {
        "summary": "Endpoint for /api/practice/details/:organizationId",
        "tags": [
          "api"
        ],
        "description": "API endpoint for /api/practice/details/:organizationId",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "organizationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "put": {
        "summary": "Endpoint for /api/practice/details/:organizationId",
        "tags": [
          "api"
        ],
        "description": "API endpoint for /api/practice/details/:organizationId",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "organizationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "summary": "Endpoint for /api/practice/details/:organizationId",
        "tags": [
          "api"
        ],
        "description": "API endpoint for /api/practice/details/:organizationId",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "organizationId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/practice/{id}": {
      "get": {
        "summary": "Endpoint for /api/practice/:id",
        "tags": [
          "api"
        ],
        "description": "API endpoint for /api/practice/:id",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "put": {
        "summary": "Endpoint for /api/practice/:id",
        "tags": [
          "api"
        ],
        "description": "API endpoint for /api/practice/:id",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "summary": "Endpoint for /api/practice/:id",
        "tags": [
          "api"
        ],
        "description": "API endpoint for /api/practice/:id",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/practice/{id}/member": {
      "post": {
        "summary": "Endpoint for /api/practice/:id/member",
        "tags": [
          "api"
        ],
        "description": "API endpoint for /api/practice/:id/member",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/practice/{id}/member/{userId}": {
      "delete": {
        "summary": "Endpoint for /api/practice/:id/member/:userId",
        "tags": [
          "api"
        ],
        "description": "API endpoint for /api/practice/:id/member/:userId",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "userId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "put": {
        "summary": "Endpoint for /api/practice/:id/member/:userId",
        "tags": [
          "api"
        ],
        "description": "API endpoint for /api/practice/:id/member/:userId",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "userId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/practice/{id}/members": {
      "get": {
        "summary": "Endpoint for /api/practice/:id/members",
        "tags": [
          "api"
        ],
        "description": "API endpoint for /api/practice/:id/members",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/practice/{id}/switch": {
      "post": {
        "summary": "Endpoint for /api/practice/:id/switch",
        "tags": [
          "api"
        ],
        "description": "API endpoint for /api/practice/:id/switch",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ]
}